split out complex constraints
